Separate commands out into modules
